# Enable Shared Memory Swizzle Optimization for Batched Matmul 