** ONLY review and DO NOT MERGE.  **

This PR uses an upstream llvm patch to enable Shared Memory Swizzle for Batched Matmul. 

`This PR`
```
matmul_128x128x768_f16t_f16t_f16t_tile_config_64x64_64x5_tensorcore_mma_sync            : 0.008 msec
BM_matmul_2x128x128x768_f16t_f16t_f16t_tile_config_64x64_64x5_tensorcore_mmasync  : 0.008 msec
```

`IREE Mainline`
```
matmul_128x128x768_f16t_f16t_f16t_tile_config_64x64_64x5_tensorcore_mma_sync            : 0.008 msec
BM_matmul_2x128x128x768_f16t_f16t_f16t_tile_config_64x64_64x5_tensorcore_mmasync  : 0.023 msec
```

Improves batched matmul performance by 2.8x for shape used for this PR's evaluation. This should also improve split-k matmul and more robust performance evaluation through IREE dispatch profiler is progress....

This PR needs an upstream LLVM patch (https://reviews.llvm.org/D148161) to be integrated into iree-llvm-fork.